requirements.txt
beautifulsoup4==4.3.2
CherryPy==3.6.0
httplib2==0.9
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.7.0
requests==2.5.2
